Open-close port per each broadcast packet
This is an ugly-but-working solution for warcraft3.
Open port (for warcraft3 it is 6112), forward broadcast
packet there and close port in order to allow the game
to get the reply packet directly. Yep, this is a race.
Nicer solution would be to create a firewall rule to
forward incoming replies to te game directly.
Printing time for race estimation. Should be smaller
than ping roundtrip.